Split error management into a separate module
